refactor: move compute_and_cache_indicators from router to service layer
Extracts the 85-line compute_and_cache_indicators function into
services/watchlist.py. Eliminates the backwards dependency where
main.py imported from a router module.

Closes #147
